+++ src/module/apps/itemImport/importer/DeviceImporter.ts
@@ -1,8 +1,7 @@
 import { DataImporter } from "./DataImporter";
 import { ImportHelper } from "../helper/ImportHelper";
 import { Constants } from "./Constants";
-import { Helpers } from "../../helpers";
-import { UpdateActionFlow } from "../../item/flows/UpdateActionFlow";
+import { UpdateActionFlow } from "../../../item/flows/UpdateActionFlow";
 
 export class DeviceImporter extends DataImporter<Shadowrun.DeviceItemData, Shadowrun.DeviceData> {
     public files = ['gear.xml'];
@@ -38,7 +37,7 @@ export class DeviceImporter extends DataImporter<Shadowrun.DeviceItemData, Shado
             item.name = ImportHelper.StringValue(commlink, 'name');
 
             // Get the item's folder information
-            //@ts-expect-error
+            // @ts-expect-error
             item.folder = folder.id;
 
             // Import Flags
@@ -86,7 +85,7 @@ export class DeviceImporter extends DataImporter<Shadowrun.DeviceItemData, Shado
             item.name = ImportHelper.StringValue(rcc, 'name');
 
             // Get the item's folder information
-            //@ts-expect-error
+            // @ts-expect-error
             item.folder = folder.id;
 
             // Import Flags
@@ -134,7 +133,7 @@ export class DeviceImporter extends DataImporter<Shadowrun.DeviceItemData, Shado
             item.name = ImportHelper.StringValue(cyberdeck, 'name');
 
             // Get the item's folder information
-            //@ts-expect-error
+            // @ts-expect-error
             item.folder = folder.id;
 
             // Import Flags
