+++ src/module/apps/itemImport/importer/EquipmentImporter.ts
@@ -1,8 +1,7 @@
 import { DataImporter } from "./DataImporter";
 import { ImportHelper } from "../helper/ImportHelper";
 import { Constants } from "./Constants";
-import { Helpers } from "../../helpers";
-import { UpdateActionFlow } from "../../item/flows/UpdateActionFlow";
+import { UpdateActionFlow } from "../../../item/flows/UpdateActionFlow";
 
 export class EquipmentImporter extends DataImporter<Shadowrun.EquipmentItemData, Shadowrun.EquipmentData> {
     files = ['gear.xml'];
@@ -56,7 +55,7 @@ export class EquipmentImporter extends DataImporter<Shadowrun.EquipmentItemData,
             // Replace / as it's used as a separator in GetFolderAtPath.
             const category = ImportHelper.TranslateCategory(categoryEN, this.categoryTranslations).replace('/', ' ');
             let categoryFolder = await ImportHelper.GetFolderAtPath(`${Constants.ROOT_IMPORT_FOLDER_NAME}/${game.i18n.localize('SR5.Gear')}/${category}`, true);
-            //@ts-expect-error
+            // @ts-expect-error
             item.folder = categoryFolder.id;
 
             // Import Flags
