+++ src/module/apps/itemImport/importer/WeaponImporter.ts
@@ -6,11 +6,10 @@ import { MeleeParser } from '../parser/weapon/MeleeParser';
 import { ThrownParser } from '../parser/weapon/ThrownParser';
 import { ParserMap } from '../parser/ParserMap';
 import { WeaponParserBase } from '../parser/weapon/WeaponParserBase';
-import { DataDefaults } from '../../data/DataDefaults';
-import { Helpers } from "../../helpers";
+import { DataDefaults } from '../../../data/DataDefaults';
 import WeaponItemData = Shadowrun.WeaponItemData;
 import WeaponData = Shadowrun.WeaponData;
-import { UpdateActionFlow } from '../../item/flows/UpdateActionFlow';
+import { UpdateActionFlow } from '../../../item/flows/UpdateActionFlow';
 
 export class WeaponImporter extends DataImporter<WeaponItemData, WeaponData> {
     public override categoryTranslations: any;
