+++ src/module/apps/itemImport/parser/critter-power/SpritePowerParser.ts
