+++ src/module/apps/itemImport/parser/spell/DetectionSpellImporter.ts
