+++ src/module/apps/itemImport/parser/spell/ManipulationSpellParser.ts
