+++ src/module/apps/itemImport/parser/weapon/ThrownParser.ts
