+++ src/module/hooks.ts
@@ -15,7 +15,7 @@ import {_combatantGetInitiativeFormula, SR5Combat} from './combat/SR5Combat';
 import {HandlebarManager} from './handlebars/HandlebarManager';
 
 import {OverwatchScoreTracker} from './apps/gmtools/OverwatchScoreTracker';
-import {Import} from './importer/apps/import-form';
+import {Import} from './apps/itemImport/apps/import-form';
 import {ChangelogApplication} from "./apps/ChangelogApplication";
 import { SituationModifiersApplication } from './apps/SituationModifiersApplication';
 import {SR5ICActorSheet} from "./actor/sheets/SR5ICActorSheet";
